+++ bot.py
@@ -2,7 +2,7 @@ from engine.network.network import *
 from engine.client import Client
 
 addr = NetAddr()
-addr.ip = bytearray("192.168.56.1".encode())
+addr.ip = bytearray("127.0.0.1".encode())
 addr.port = 8303
